% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{access}
\alias{access}
\title{access variable from report}
\usage{
access(report, name)
}
\arguments{
\item{report}{Report object}

\item{name}{variable name}
}
\value{
variable data
}
\description{
access variable from report
}
