% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{add}
\alias{add}
\title{add variable to report}
\usage{
add(report, name, value)
}
\arguments{
\item{report}{Report object}

\item{name}{variable name}

\item{value}{variable value}
}
\value{
Report object
}
\description{
add variable to report
}
