% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{as.character.FileBackend}
\alias{as.character.FileBackend}
\title{convert FileBackend to character}
\usage{
\method{as.character}{FileBackend}(x, ...)
}
\arguments{
\item{x}{FileBackend object}

\item{...}{refer to as.character for details.}
}
\value{
FileBackend character
}
\description{
convert FileBackend to character
}
