% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{as.character.Report}
\alias{as.character.Report}
\title{convert Report object to character}
\usage{
\method{as.character}{Report}(x, ...)
}
\arguments{
\item{x}{Report object}

\item{...}{refer to as.character for details.}
}
\description{
convert Report object to character
}
