% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{FileBackend}
\alias{FileBackend}
\title{FileBackend}
\usage{
FileBackend(reporter_dir = NULL)
}
\arguments{
\item{reporter_dir}{a directory that save report data}
}
\value{
FileBackend object
}
\description{
FileBackend
}
