% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{get_report_file}
\alias{get_report_file}
\title{get report file for FileBackend}
\usage{
get_report_file(backend, report_id)
}
\arguments{
\item{backend}{FileBackend object}

\item{report_id}{report UUID}
}
\value{
report file path
}
\description{
get report file for FileBackend
}
