% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{read.FileBackend}
\alias{read.FileBackend}
\title{read report from FileBackend}
\usage{
\method{read}{FileBackend}(backend, report_id, ...)
}
\arguments{
\item{backend}{FileBackend object}

\item{report_id}{report UUID}

\item{...}{refer to read for details.}
}
\value{
report data
}
\description{
read report from FileBackend
}
