% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{read.Report}
\alias{read.Report}
\title{read data from report}
\usage{
\method{read}{Report}(report, name, ...)
}
\arguments{
\item{report}{Report object}

\item{name}{variable name}

\item{...}{refer to read for details.}
}
\description{
read data from report
}
