% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{Report}
\alias{Report}
\title{Report}
\usage{
Report(report_id = NULL, backend = NULL)
}
\arguments{
\item{report_id}{UUID string or NULL}

\item{backend}{Backend object}
}
\value{
Report object
}
\description{
Report
}
