% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{store.FileBackend}
\alias{store.FileBackend}
\title{store report to FileBackend}
\usage{
\method{store}{FileBackend}(backend, report_id, data, ...)
}
\arguments{
\item{backend}{FileBackend object}

\item{report_id}{an UUID string}

\item{data}{list object}

\item{...}{refer to store for details}
}
\value{
file path
}
\description{
store report to FileBackend
}
