% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{store.Report}
\alias{store.Report}
\title{save report to backend}
\usage{
\method{store}{Report}(report, ...)
}
\arguments{
\item{report}{Report object}

\item{...}{refer to store for details.}
}
\description{
save report to backend
}
